requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers

chromadb
langchain-groq
python-dotenv
google-generativeai
langchain-google-genai
fastapi
uvicorn